reorganise the add_guest step to encapsulate everything for that step
