+++ add_guest/__init__.py
@@ -1,0 +1,2 @@
+# flake8: noqa
+from .add_guest import routes
