+++ views/__init__.py
@@ -2,4 +2,3 @@
 from .home import Home
 from .list_invites import ListInvites
 from .create_invite import CreateInvite
-from .add_guest import AddGuest
